requirements.txt
django==3.2.*
environs==11.0.0
django-admin-sortable2==2.2.1